clients/banchi-client/pyproject.toml
[tool.poetry]
name = "banchi-client"
version = "0.0.2"
description = "A client library for accessing Banchi"
authors = []
readme = "README.md"
packages = [
    { include = "banchi_client" },
]
include = ["banchi_client/py.typed"]

[tool.poetry.dependencies]
python = "^3.10"
httpx = ">=0.23.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
